Initial implementation of 'readonly' plugin
The initial implementation is based on the 'readonly' groovy script
from the 'scripts' project. The advantage of splitting it out to a
standalone full plugin is that it is not necessary for the groovy
scripting plugin to also be installed on the server.

The functionality implemented in the initial version is the same as
in the groovy script, and also has the same limitations. That is, it
can only prevent changes from being made to repositories by git push.
It does not make Gerrit truly read-only because is still possible to
cause server-side changes to be made in various other ways:

- Creating a new project
- Changing a project configuration (without review)
- Any PUT/POST/DELETE request via the REST API
- Adding review comments via SSH command
- and possibly others

Prevention of these operations should be implemented in follow-up
commits.

Change-Id: If9a5ec53ec9d33001fa8fd32c895f1448dc26347