Add read only HTTP filter to prevent PUT/POST/DELETE
The filter intercepts all HTTP requests and rejects any PUT, POST
or DELETE with a "503 Service Unavailable" status code [1].

This prevents:

- Creating a new project from the UI
- Changing a project's configuration from the UI
- Changing user preferences from the UI
- and any other operation performed by a POST/PUT/DELETE REST API call

A side effect of this is that login is also disabled, since it relies
on a POST request. Whether to allow logins should be discussed and
implemented in a follow-up change if necessary.

[1] https://httpstatuses.com/503

Change-Id: I4fa7d335d8392270e004f55d29581eeffa3346a2