Tighten up restriction on plugin ssh commands
The plugin currently allows all "gerrit plugin" commands, which means
it is possible to install a new plugin while Gerrit is in read-only
mode.

Make it so the only commands allowed are "rm" and "remove" for this
plugin. This will ensure that no other plugins may be reloaded or
added while the readonly plugin is active.

Change-Id: Ie6bb5584ec4559bb8434706f571b86f9ff718776