Update patch set 2
Patch Set 2: -Code-Review

Mocked based tests is removed here. The whole build tool chain, that was extended in on of the parent change should be reverted.

A different approach would be to revert all mocked based tests and the Bazel build extension in the parent change. Yet another alternative is to remove it in follow-up change. But we shouldn't leave dead code when this series is merged.

Patch-set: 2
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0