Update patch set 5
Patch Set 5:

(1 comment)

The code looks OK, but I wonder if we should do this with js rather than GWT as Jacek suggested.

If we do it with js will it be compatible out of the box with polygerrit?  Or would it require extra work?

Patch-set: 5
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>