Http: fix generation of the custom readonly message
The use of custom HTTP status line is deprecated and has been removed
in Jetty since v9.4.21 (see [1]). The custom message when Gerrit
is read-only must be provided as response body.

Adapt the Git/HTTP tests accordingly, as JGit does not read the response
body if the status code is not 200 and therefore there is no way
to get the custom message anymore over the Git protocol.

[1]: https://github.com/eclipse/jetty.project/issues/4154

Bug: Issue 12146
Change-Id: I6ae10605fefa470d283492bb85147332f1438d0d