Add option to configure location of marker file
In a setup using the high-availability plugin it can be useful to
put all instances into read-only mode with a single command. To do this
the marker file has to be in the shared directory.

This change allows to configure the directory in which the marker file
will be created and watched. This allows to configure for example the
shared directory in an HA-setup.

Change-Id: I492b2141c697bbe46edf106800d8dfe42275b965